requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
shap
tensorflow-cpu
joblib